requirements.txt
SQLAlchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
fastapi==0.109.2
aiosqlite==0.20.0
uvicorn==0.27.1
Jinja2==3.1.3
sqladmin==0.16.1


Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
python-dotenv==0.19.0
